requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
construct==2.10.57
discord.py==1.6.0
exiles-api>=1.16.19
factorio-rcon-py==1.2.1
idna==3.1
multidict==5.1.0
psutil==5.8.0
PySocks==1.7.1
python-dotenv==0.15.0
SQLAlchemy==1.3.22
typing-extensions==3.7.4.3
yarl==1.6.3
